Reorganize modules to avoid context problems
These kind of changes are necessary to dodge
problems that stem from LIBRARY AUTHORS RELYING
ON INTERAL STATE AND EFFECTS, in this case i am
referring to mongoose, the community knows it is
a problem, but the general contributor consensus
has been that the effort required to refactor the
problem away is enormous and the workaround to
avoid the problem is simple.

I think they should fix it, because it causes
very confusing errors that are difficult to
resolve without cracking open the source code
and tracing the path of every call, and the point
of shared libraries is to NOT have to do that.